LF-4078 - Merge updates separate out pure component smoall type fixes some type errors still exsist on themes and table props
