+++ packages/webapp/src/components/Finances/Transaction/ExpandedContent/CropSaleTable.jsx
@@ -15,6 +15,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import Table from '../../../Table/Table';
+import { TableType } from '../../../Table/types';
 import history from '../../../../history';
 import styles from './styles.module.scss';
 import { createRevenueDetailsUrl } from '../../../../util/siteMapConstants';
@@ -83,7 +84,7 @@ export default function CropSaleTable({ data, currencySymbol, mobileView }) {
 
   return (
     <Table
-      kind="v2"
+      kind={TableType.V2}
       columns={getColumns(t, mobileView, totalAmount, quantityWithUnit, currencySymbol)}
       data={items}
       minRows={10}
