+++ packages/webapp/src/components/Finances/Transaction/ExpandedContent/GeneralTransactionTable.jsx
@@ -15,6 +15,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import Table from '../../../Table/Table';
+import { TableType } from '../../../Table/types';
 import { transactionTypeEnum } from '../../../../containers/Finances/useTransactions';
 
 const formatData = (t, data) => {
@@ -64,7 +65,7 @@ export default function GeneralTransactionTable({ data, currencySymbol, mobileVi
 
   return (
     <Table
-      kind="v2"
+      kind={TableType.V2}
       columns={getColumns(mobileView, currencySymbol, titleLabel, amountLabel)}
       data={tableData}
       shouldFixTableLayout={true}
