+++ packages/webapp/src/components/Profile/People/index.jsx
@@ -4,6 +4,7 @@ import React, { useState } from 'react';
 import Button from '../../Form/Button';
 import PropTypes from 'prop-types';
 import Table from '../../Table/Table';
+import { TableType } from '../../Table/types';
 import ProfileLayout from '../ProfileLayout';
 
 export default function PurePeople({ users, history, isAdmin }) {
@@ -122,7 +123,7 @@ export default function PurePeople({ users, history, isAdmin }) {
         placeholder={t('PROFILE.PEOPLE.SEARCH')}
       />
       <Table
-        kind="v1"
+        kind={TableType.V1}
         data-cy="people-table"
         columns={summaryColumns}
         data={getFilteredUsers()}
