+++ packages/webapp/src/containers/Finances/Labour/Employee/index.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import Table from '../../../../components/Table/Table';
+import { TableType } from '../../../../components/Table/types';
 import moment from 'moment';
 import { useTranslation } from 'react-i18next';
 import { userFarmsByFarmSelector } from '../../../userFarmSlice';
@@ -54,7 +55,7 @@ const Employee = ({ currencySymbol, tasks, startDate, endDate }) => {
 
   return (
     <Table
-      kind="v1"
+      kind={TableType.V1}
       columns={columns}
       data={data}
       showPagination={true}
