+++ packages/webapp/src/containers/Finances/OtherExpense/index.jsx
@@ -20,6 +20,7 @@ import defaultStyles from '../styles.module.scss';
 import styles from './styles.module.scss';
 import { expenseSelector, allExpenseTypeSelector } from '../selectors';
 import Table from '../../../components/Table/Table';
+import { TableType } from '../../../components/Table/types';
 import { getExpense, getFarmExpenseType } from '../actions';
 import history from '../../../history';
 import DateRangeSelector from '../../../components/Finances/DateRangeSelector';
@@ -193,7 +194,7 @@ const OtherExpense = () => {
       <div className={styles.tableContainer} style={{ marginBottom: '16px' }}>
         {data.length > 0 && (
           <Table
-            kind="v1"
+            kind={TableType.V1}
             columns={columns}
             data={data}
             showPagination={true}
@@ -210,7 +211,7 @@ const OtherExpense = () => {
         {detailedHistory.length > 0 && (
           <div>
             <Table
-              kind="v1"
+              kind={TableType.V1}
               columns={detailedColumns}
               data={detailedHistory}
               showPagination={true}
